requirements.txt
blinker==1.9.0
click==8.1.7
Flask==2.3.3
Flask-Mail==0.9.1
itsdangerous==2.1.1
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0